Update for tcpdump

+++ scripts/tcpdump.sh
@@ -3,11 +3,12 @@
 set -e
 set -x 
 
-declare -ir file_count=${1:? error ... file_count: <NUMBER>} 
-declare -ir file_size=${2:? error ... file_size: <NUMBER>}
+if ! [[ -f ${0%.sh}.env ]]; then
+    echo ${0%.sh}.env not found
+    exit 1
+fi
 
-declare -r dump_file_name=tcpdump-$(date +%s)
-declare -r inet_name=$(ip route show default | cut -f5 -d ' ')
+source $PWD/${0%.sh}.env
 
 tcpdump -i $inet_name -w ${dump_file_name}.pcap. -W $file_count -C $file_size -K -n >> $PWD/${dump_file_name}.log 2>&1 &
 echo ${1} > ${dump_file_name}.pid
